Started commentSectionManager.js
Changed comment printing by php so that every comment has a id 'c?' where ? is the number of it. Also changed the sql query so comments are visible in order - most recent on the top. CommentSectionManager script introduced. For now grabs all comment objects and splits them into the commentsHidable array - these one are going to be interrupted by the function showHideComments() in the future.